Bad parameter type in write8 prototype.

+++ Adafruit_TSL2561_U.h
@@ -194,7 +194,7 @@ class Adafruit_TSL2561_Unified : public Adafruit_Sensor {
   
   void     enable (void);
   void     disable (void);
-  void     write8 (uint8_t reg, uint32_t value);
+  void     write8 (uint8_t reg, uint8_t value);
   uint8_t  read8 (uint8_t reg);
   uint16_t read16 (uint8_t reg);
   void     getData (uint16_t *broadband, uint16_t *ir);
